+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added the `jwt:generate` command to generate keys for signing JWTs ([#7])
 
 ## [1.1.2] - 2020-01-03
 ### Fixed
@@ -38,3 +40,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.1.0]: https://github.com/sprocketbox/laravel-jwt/compare/v1.0...v1.1.0
 [1.0.0]: https://github.com/sprocketbox/laravel-jwt/releases/tag/v1.0
 [#5]: https://github.com/sprocketbox/laravel-jwt/pull/5
+[#7]: https://github.com/sprocketbox/laravel-jwt/issues/7